requirements.txt
Flask==2.3.3
langchain-text-splitters==0.3.8
langchain-community==0.3.21
langchain-core==0.3.51
langchain-chroma==0.2.2
chromadb==0.4.22
langchain-huggingface==0.1.2
sentence-transformers==2.5.1
langchain-groq==0.2.0
python-dotenv==1.0.1
Werkzeug==2.3.7 